+++ plugins/ui/src/deephaven/ui/__init__.py
@@ -9,4 +9,4 @@ from .elements import *
 from .hooks import *
 from .object_types import *
 
-__version__ = "0.7.0.dev0"
+__version__ = "0.8.0.dev0"