+++ plugins/ui/src/deephaven/ui/components/__init__.py
@@ -8,6 +8,9 @@ from .dashboard import dashboard
 from .row import row
 from .column import column
 from .stack import stack
+from .picker import picker
+from .section import section
+from .item import item
 
 from . import html
 
@@ -29,13 +32,15 @@ __all__ = [
     "heading",
     "icon",
     "icon_wrapper",
+    "item",
     "illustrated_message",
     "html",
     "number_field",
-    "item",
     "panel",
+    "picker",
     "range_slider",
     "row",
+    "section",
     "slider",
     "spectrum_element",
     "stack",