+++ plugins/ui/test/deephaven/ui/test_hooks.py
@@ -108,7 +108,7 @@ class HooksTest(BaseTestCase):
         from deephaven.ui.hooks import use_memo
 
         def _test_memo(fn=lambda: "foo", a=1, b=2):
-            return use_memo(fn, [a, b])
+            return use_memo(fn, {a, b})
 
         # Initial render
         render_result = render_hook(_test_memo)
@@ -151,7 +151,7 @@ class HooksTest(BaseTestCase):
             event.set()
 
         def _test_table_listener(replayed_table_val=table, listener_val=listener):
-            use_table_listener(replayed_table_val, listener_val)
+            use_table_listener(replayed_table_val, listener_val, [])
 
         render_hook(_test_table_listener)
 
@@ -171,7 +171,7 @@ class HooksTest(BaseTestCase):
             event.set()
 
         def _test_table_listener(replay_table=table, listener_val=listener):
-            use_table_listener(replay_table, listener_val, do_replay=True)
+            use_table_listener(replay_table, listener_val, [], do_replay=True)
 
         render_hook(_test_table_listener)
 
@@ -541,7 +541,7 @@ class HooksTest(BaseTestCase):
                 thread.start()
                 thread.join()
 
-            use_memo(start_thread, [])
+            use_memo(start_thread, set())
 
         render_hook(_test_execution_context)
 
@@ -564,7 +564,7 @@ class HooksTest(BaseTestCase):
                 thread = threading.Thread(target=thread_func)
                 thread.start()
 
-            use_memo(start_thread, [])
+            use_memo(start_thread, set())
 
         render_hook(_test_execution_context)
 
@@ -595,7 +595,7 @@ class HooksTest(BaseTestCase):
             a, set_a = use_state(lambda: table.where("X=1"))
 
             # When "a" changes, recompute table - don't return or otherwise track this table w.r.t. liveness
-            replace_a = use_liveness_scope(lambda: set_a(table.where("X=2")))
+            replace_a = use_liveness_scope(lambda: set_a(table.where("X=2")), set())
 
             return a.size